Title: 올바른 등식 만들기, Time: 6ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 240806/올바른 등식 만들기/right-equality.cpp
@@ -5,7 +5,7 @@ using namespace std;
 
 int n,m;
 int arr[101];
-int dp[101][41];
+long long dp[101][41];
 
 int main() {
     cin>>n>>m;
